- Move Register form and components to components/forms - Move Register methods to helper file
