+++ app/components/helpers/auth.ts
@@ -17,3 +17,20 @@ export const validateEmail = (email: unknown): email is string => {
 export const validatePassword = (password: unknown): password is string => {
 	return typeof password === 'string' && password.length >= 8 && password.length <= 255;
 }
+
+export const validateFormData = (firstName: string, lastName: string, email: string, password: string, confirmPassword: string) => {
+	if (!firstName && !lastName && !email && !password && !confirmPassword) {
+		return false;
+	} else if (!validateFormComplete(firstName, email, password, confirmPassword)) {
+		return false;
+	} else if (!validateEmail(email)) {
+		return false
+	} else if (!validatePassword(password)) {
+		return false
+	} else if (password !== confirmPassword) {
+		return false
+	} else {
+		return true
+	}
+}
+
